// src/types/anilist.ts
// Generated by https://quicktype.io
import * as Main from "./main";

export interface AnilistRes {
	data: AnilistResData;
}

export interface AnilistResData {
	MediaListCollection: MediaListCollection;
}

export interface FormattedAnilistResData {
	stats: Main.Stats;
	list: Array<Main.Media>;
}

export interface MediaListCollection {
	lists: List[];
}

export interface List {
	entries: Entry[];
	name: string;
	isCustomList: boolean;
	isSplitCompletedList: boolean;
	status: string;
}

export interface Entry {
	id: number;
	status: string;
	score: number;
	progress: number;
	notes: string | null;
	repeat: number;
	media: Media;
}

export interface Media {
	chapters: number | null;
	volumes: number | null;
	idMal: number | null;
	episodes: number | null;
	title: Title;
}

export interface Title {
	romaji: string;
}